WIP: Volumetric path tracing
