+++ renderer/vpt.py
@@ -1,0 +1,26 @@
+"""
+    Volumetric Path Tracer
+    Participating media is bounded by object or world AABB
+    @author: Qianyue He
+    @date: 2023-2-13
+"""
+
+import taichi as ti
+from taichi.math import vec3
+
+from typing import List
+from la.cam_transform import *
+from tracer.path_tracer import PathTracer
+from emitters.abtract_source import LightSource
+
+from scene.obj_desc import ObjDescriptor
+from sampler.general_sampling import mis_weight
+
+@ti.data_oriented
+class VolumeRenderer(PathTracer):
+    """
+        Volumetric Renderer Final Class
+    """
+    def __init__(self, emitters: List[LightSource], objects: List[ObjDescriptor], prop: dict):
+        super().__init__(emitters, objects, prop)
+
