+++ tracer/path_tracer.py
@@ -36,7 +36,7 @@ class PathTracer(TracerBase):
         """
             Implement path tracing algorithms first, then we can improve light source / BSDF / participating media
         """
-        timer = TicToc()
+        self.clock = TicToc()
         self.anti_alias         = prop['anti_alias']
         self.stratified_sample  = prop['stratified_sampling']   # whether to use stratified sampling
         self.use_mis            = prop['use_mis']               # whether to use multiple importance sampling
@@ -59,10 +59,11 @@ class PathTracer(TracerBase):
         self.brdf_nodes.place(self.brdf_field)                              # BRDF Taichi storage
         ti.root.bitmasked(ti.i, self.num_objects).place(self.bsdf_field)    # BRDF Taichi storage (no node needed)
 
-        print(f"[INFO] Path tracer param loading in {timer.toc(True):.3f} ms")
-        timer.tic()
+        print(f"[INFO] Path tracer param loading in {self.clock.toc(True):.3f} ms")
+        self.clock.tic()
         self.initialze(emitters, objects)
-        print(f"[INFO] Path tracer initialization in {timer.toc(True):.3f} ms")
+        print(f"[INFO] Path tracer initialization in {self.clock.toc(True):.3f} ms")
+        self.clock.tic()
 
     def initialze(self, emitters: List[LightSource], objects: List[ObjDescriptor]):
         for i, emitter in enumerate(emitters):
